refactor: remove warnings, change result file, add timestamp to results

+++ lib/graph.hpp
@@ -25,7 +25,7 @@ private:
 
 class ThreadData {
 public:
-  int k;
+  long unsigned int k;
   set<vector<int>> cliques;
   Graph *graph;
 };
